+++ layer.py
@@ -166,6 +166,13 @@ class BaseLayer(QgsVectorLayer):
             raise IOError(msg)
         return writer
 
+    @staticmethod
+    def delete_shp(path):
+        QgsVectorFileWriter.deleteShapeFile(path)
+        path = os.path.splitext(path)[0] + '.cpg'
+        if os.path.exists(path):
+            os.remove(path)
+
     def copy_feature(self, feature, rename=None, resolve=None):
         """
         Return a copy of feature renaming attributes or resolving xlink references.
@@ -993,6 +1000,7 @@ class AddressLayer(BaseLayer):
                 QgsField('localId', QVariant.String, len=254),
                 QgsField('spec', QVariant.String, len=254),
                 QgsField('designator', QVariant.String, len=254),
+                QgsField('image', QVariant.String, len=254),
                 QgsField('PD_id', QVariant.String, len=254),
                 QgsField('TN_id', QVariant.String, len=254),
                 QgsField('AU_id', QVariant.String, len=254)
@@ -1059,6 +1067,14 @@ class AddressLayer(BaseLayer):
                 highway_names[name] = hgwnames.match(name, choices)
         return highway_names
 
+    def get_image_links(self):
+        to_change = {}
+        for feat in self.getFeatures():
+            url = setup.cadastre_doc_url.format(feat['localId'][-14:])
+            feat['image'] = url
+            to_change[feat.id()] = get_attributes(feat)
+        self.writer.changeAttributeValues(to_change)
+
 
 class ConsLayer(PolygonLayer):
     """Class for constructions"""
@@ -1070,7 +1086,7 @@ class ConsLayer(PolygonLayer):
             self.writer.addAttributes([
                 QgsField('localId', QVariant.String, len=254),
                 QgsField('condition', QVariant.String, len=254),
-                QgsField('link', QVariant.String, len=254),
+                QgsField('image', QVariant.String, len=254),
                 QgsField('currentUse', QVariant.String, len=254),
                 QgsField('bu_units', QVariant.Int),
                 QgsField('dwellings', QVariant.Int),
@@ -1083,7 +1099,7 @@ class ConsLayer(PolygonLayer):
             self.updateFields()
         self.rename = {
             'condition': 'conditionOfConstruction',
-            'link': 'documentLink' ,
+            'image': 'documentLink' ,
             'bu_units': 'numberOfBuildingUnits',
             'dwellings': 'numberOfDwellings',
             'lev_above': 'numberOfFloorsAboveGround',