+++ test/test_layer.py
@@ -150,7 +150,8 @@ class TestBaseLayer(unittest.TestCase):
         self.layer.updateFields()
 
     def tearDown(self):
-        QgsVectorFileWriter.deleteShapeFile('test_layer.shp')
+        del self.layer
+        BaseLayer.delete_shp('test_layer.shp')
 
     def test_copy_feature_with_resolve(self):
         feature = self.fixture.getFeatures().next()
@@ -289,24 +290,28 @@ class TestBaseLayer(unittest.TestCase):
         layer.test(layer)
         m_index.assert_called_with(layer.getFeatures.return_value)
 
+class TestBaseLayer2(unittest.TestCase):
+
     @mock.patch('layer.QgsVectorFileWriter')
     @mock.patch('layer.os')
     def test_export_default(self, mock_os, mock_fw):
+        layer = BaseLayer("Polygon", "test", "memory")
         mock_os.path.exists.side_effect = lambda arg: arg=='foobar'
         mock_fw.writeAsVectorFormat.return_value = QgsVectorFileWriter.NoError
         mock_fw.NoError = QgsVectorFileWriter.NoError
-        self.assertTrue(self.layer.export('foobar'))
+        self.assertTrue(layer.export('foobar'))
         mock_fw.deleteShapeFile.assert_called_once_with('foobar')
-        mock_fw.writeAsVectorFormat.assert_called_once_with(self.layer, 'foobar',
-            'utf-8', self.layer.crs(), 'ESRI Shapefile')
+        mock_fw.writeAsVectorFormat.assert_called_once_with(layer, 'foobar',
+            'utf-8', layer.crs(), 'ESRI Shapefile')
 
     @mock.patch('layer.QgsVectorFileWriter')
     @mock.patch('layer.os')
     def test_export_other(self, mock_os, mock_fw):
+        layer = BaseLayer("Polygon", "test", "memory")
         mock_os.path.exists.side_effect = lambda arg: arg=='foobar'
-        self.layer.export('foobar', 'foo')
+        layer.export('foobar', 'foo')
         mock_os.remove.assert_called_once_with('foobar')
-        self.layer.export('foobar', 'foo', overwrite=False)
+        layer.export('foobar', 'foo', overwrite=False)
         mock_os.remove.assert_called_once_with('foobar')
 
 
@@ -323,7 +328,8 @@ class TestPolygonLayer(unittest.TestCase):
         self.assertEquals(self.layer.featureCount(), self.fixture.featureCount())
 
     def tearDown(self):
-        QgsVectorFileWriter.deleteShapeFile('test_layer.shp')
+        del self.layer
+        PolygonLayer.delete_shp('test_layer.shp')
 
     def test_get_multipolygon(self):
         p = [[QgsPoint(0,0), QgsPoint(1,0), QgsPoint(1,1), QgsPoint(0,0)]]
@@ -461,6 +467,12 @@ class TestZoningLayer(unittest.TestCase):
             g = f.geometry()
             self.assertFalse(g.isMultipart())
 
+    def tearDown(self):
+        del self.layer1
+        ZoningLayer.delete_shp('urban_zoning.shp')
+        del self.layer2
+        ZoningLayer.delete_shp('rustic_zoning.shp')
+
     def test_get_adjacents_and_geometries(self):
         (groups, geometries) = self.layer1.get_adjacents_and_geometries()
         self.assertTrue(all([len(g) > 1 for g in groups]))
@@ -520,7 +532,8 @@ class TestConsLayer(unittest.TestCase):
         self.assertEquals(self.layer.featureCount(), self.fixture.featureCount())
 
     def tearDown(self):
-        QgsVectorFileWriter.deleteShapeFile('test_layer.shp')
+        del self.layer
+        ConsLayer.delete_shp('test_layer.shp')
 
     def test_is_building(self):
         self.assertTrue(ConsLayer.is_building({'localId': 'foobar'}))
@@ -996,7 +1009,8 @@ class TestAddressLayer(unittest.TestCase):
         self.layer.updateFields()
 
     def tearDown(self):
-        QgsVectorFileWriter.deleteShapeFile('test_layer.shp')
+        del self.layer
+        AddressLayer.delete_shp('test_layer.shp')
 
     def test_append(self):
         self.layer.append(self.address_gml)