+++ test/test_osmxml.py
@@ -35,20 +35,22 @@ class OsmxmlTest(unittest.TestCase):
         for (xmltag, osmtag) in zip(root.findall('node/tag'), n.tags.items()):
             self.assertEquals(xmltag.get('k'), osmtag[0])
             self.assertEquals(xmltag.get('v'), osmtag[1])
-        self.assertEquals(root.xpath('count(//way)'), 3)
+        nw = 0
         for (xmlway, osmway) in zip(root.findall('way'), data.ways):
+            nw += 1
             for (xmlnd, osmnd) in zip(xmlway.findall('nd'), osmway.nodes):
                 self.assertEquals(int(xmlnd.get('ref')), osmnd.id)
+        self.assertEquals(nw, 3)
         for (xmltag, osmtag) in zip(root.findall('way/tag'), w.tags.items()):
             self.assertEquals(xmltag.get('k'), osmtag[0])
             self.assertEquals(xmltag.get('v'), osmtag[1])
-        self.assertEquals(root.xpath('count(//relation)'), 1)
         for (i, (xmlm, osmm)) in enumerate(zip(root.findall('relation/member'), r.members)):
             self.assertEquals(int(xmlm.get('ref')), osmm.ref)
             self.assertEquals(xmlm.get('role'), 'outer' if i == 0 else 'inner')
         for (xmltag, osmtag) in zip(root.findall('relation/tag'), r.tags.items()):
             self.assertEquals(xmltag.get('k'), osmtag[0])
             self.assertEquals(xmltag.get('v'), osmtag[1])
+        self.assertEquals(sum([1 for r in root.findall('relation')]), 1)
         self.assertEquals(root.find('note'), None)
         self.assertEquals(root.find('meta'), None)
         data.note = 'foobar'