+++ test/test_setup.py
@@ -9,6 +9,7 @@ import setup
 class TestSetup(unittest.TestCase):
 
     def test_win(self):
+        eol = setup.eol
         lang = os.getenv('LANG')
         setup.platform = 'linux2'
         setup.winenv()
@@ -21,4 +22,5 @@ class TestSetup(unittest.TestCase):
         setup.winenv()
         self.assertEquals(os.getenv('LANG'), 'foobar')
         os.environ['LANG'] = lang
+        setup.eol = eol